fix comparison

+++ src/ray/common/cgroup/cgroup_setup.cc
@@ -284,7 +284,8 @@ Status CgroupSetup::AddInternalProcess(pid_t pid) {
 
 ScopedCgroupHandler CgroupSetup::ApplyCgroupForDefaultAppCgroup(
     const AppProcCgroupMetadata &ctx) {
-  RAY_CHECK_EQ(ctx.max_memory, 0) << "Ray doesn't support per-task resource constraint.";
+  RAY_CHECK_EQ(ctx.max_memory, static_cast<uint64_t>(0))
+      << "Ray doesn't support per-task resource constraint.";
 
   std::ofstream out_file(cgroup_v2_default_app_proc_filepath_,
                          std::ios::app | std::ios::out);
